refactor: standardize theme unit test file locations
- Move nord.test.ts and lumos.test.ts into lib/svg/themes/ directory
- Consistent theme test organization across the repository
- All individual theme tests now located in lib/svg/themes/
